preparing sve top bits
