+++ include/eve/detail/function/friends.hpp
@@ -21,3 +21,7 @@
 #if defined(EVE_INCLUDE_ARM_HEADER)
 #  include <eve/detail/function/simd/arm/neon/friends.hpp>
 #endif
+
+#if defined(EVE_INCLUDE_SVE_HEADER)
+#  include <eve/detail/function/simd/arm/sve/friends.hpp>
+#endif
